requirements.txt
streamlit
python-dotenv
hume[microphone]

aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
eval-type-backport==0.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
simpleaudio==1.0.4
sniffio==1.3.1
sounddevice==0.4.7
typing-extensions==4.12.2
websockets==12.0